requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu

torch 
torchvision 
torchaudio

argparse==1.4.0
aiocontextvars==0.2.2
cython==0.29.23
msgpack-numpy==0.4.8
msgpack==1.0.2
python-json-logger==2.0.2
pyyaml==6.0.0
fastapi==0.95.0
uvicorn==0.21.1
uvloop==0.17.0
psutil==5.9.4
joblib==1.2.0
requests==2.31.0
huggingface-hub==0.16.4

sentence-transformers==2.2.2